Establish database connection with mongoose
The code establishes a connection to the MongoDB database specified by the MONGO_URI environment variable. It utilizes the provided ConnectOptions to configure the connection settings, including the use of useNewUrlParser and useUnifiedTopology. Upon successful connection, it logs the details of the connected database. If there is an error during the connection process, it logs the error message